new module in config dedicated to initialize host configuration files
