+++ tools/config/src/lib.rs
@@ -1,3 +1,5 @@
 pub mod config_cli;
 pub mod legacy_config;
 pub mod make_profile;
+pub mod host;
+pub mod node_yaml;
